Update `artichoke` CLI to clap v4
https://epage.github.io/blog/2022/09/clap4/

`--help` output is changed, so update the ui-test snapshots.
