+++ Cargo.toml
@@ -19,7 +19,7 @@ homepage.workspace = true
 documentation.workspace = true
 
 [dependencies]
-clap = { version = "3.2.5", optional = true, default-features = false, features = ["std", "suggestions"] }
+clap = { version = "4.0.2", optional = true }
 # XXX: load-bearing unused dependency.
 #
 # `rustyline` improperly declares its minimum version on `log` as `0.4` despite
