+++ README.md
@@ -94,22 +94,20 @@ temporary workaround is to inject data into the interpreter with the
 
 ```console
 $ artichoke --help
-artichoke 0.1.0-pre.0
 Artichoke is a Ruby made with Rust.
 
-USAGE:
-    artichoke [OPTIONS] [ARGS]
+Usage: artichoke [OPTIONS] [programfile] [arguments]...
 
-ARGS:
-    <programfile>
-    <arguments>...
+Arguments:
+  [programfile]
+  [arguments]...
 
-OPTIONS:
-        --copyright                 print the copyright
-    -e <commands>                   one line of script. Several -e's allowed. Omit [programfile]
-    -h, --help                      Print help information
-    -V, --version                   Print version information
-        --with-fixture <fixture>    file whose contents will be read into the `$fixture` global
+Options:
+      --copyright               print the copyright
+  -e <commands>                 one line of script. Several -e's allowed. Omit [programfile]
+      --with-fixture <fixture>  file whose contents will be read into the `$fixture` global
+  -h, --help                    Print help information
+  -V, --version                 Print version information
 ```
 
 ## Design and Goals
